% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.R
\name{accumulate}
\alias{accumulate}
\title{Cumulative Sums, Product, and Extremes by Subset}
\usage{
accumulate(formula, data, FUN = cumsum, rev = FALSE)
}
\arguments{
\item{formula}{An object of class \code{formula} that symbolizes how
elements should be group and which dependent element should be
accumulated.}

\item{data}{A data frame containing the variables in the formula.}

\item{FUN}{Function to apply to each group.}

\item{rev}{Whether or not to perform \code{FUN} on a reversed
ordering of elements in \code{data}.}
}
\value{
A vector of accumulated values
}
\description{
Returns a vector whose elements are the cumulative sums, products,
minima, or maxima of the elements in data frame, according to a
specified grouping.
}
\examples{
df <- data.frame(X1 = c('a', 'b', 'b', 'b', 'a', 'a', 'b'),
                 X2 = c('c', 'c', 'd', 'c', 'd', 'd', 'c'),
                 C1 = c(2, 5, 1, 7, 3, 1, 2))
accumulate(C1 ~ X1 + X2, df, FUN = cumsum)
}

