% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alarm.R
\name{alarm}
\alias{alarm}
\title{Play an alarm}
\usage{
alarm(x)
}
\arguments{
\item{x}{Location of the sound file relative the path specified in
\code{ben.options('path.alarms')}}
}
\description{
Plays a sound file using Powershell.
}

