% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_correlations.R
\name{average_correlations}
\alias{average_correlations}
\title{Average Correlations with Fisher's R to Z conversion}
\usage{
average_correlations(correlations, ns, na.rm = T, use_weights = F)
}
\arguments{
\item{correlations}{The vector of correlations to average}
}
\value{
float: Average of correlations
}
\description{
This function will take as input a verctor of correlations and output their mean after being transformed to fisher's Zs and back
}
\details{
see https://www.tandfonline.com/doi/pdf/10.1080/00221309809595548?needAccess=true
}
\examples{
c(.40, .04, .23, .25, .94, .01) |> average_correlations()
c(.40, .04, .23, .25, .94, .01) |> average_correlations(ns = c(10, 20, 40, 23, 10, 123))
c(.40, .04, .23, .25, .94, .01) |> average_correlations(ns = c(10, 20, 40, 23, 10, 123), use_weights = F)
c(.40, .04, .23, .25, .94, .01) |> average_correlations(ns = c(10, 20, 40, 23, 10, 123), use_weights = T)
}
