% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_excel.R
\name{bind_excel}
\alias{bind_excel}
\title{Combine Excel Sheets by Rows}
\usage{
bind_excel(path, pattern = NULL, progress = interactive())
}
\arguments{
\item{path}{Vector of file paths representing the Excel documents to
be combined.}

\item{pattern}{Regular expression to be matched to sheets that should
be combined.}

\item{progress}{Whether to print verbose messages on the progress of
the method. By default this will only display in an interactive
session.}
}
\value{
A tibble.
}
\description{
Take a vector of Excel workbook paths and combine all of the
component worksheets by rows.
}
