% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_convert.R
\name{c2numeric}
\alias{c2numeric}
\title{Convert a currency-formatted string to a numeric}
\usage{
c2numeric(x, dm = ".")
}
\arguments{
\item{x}{Character vector to be converted}

\item{dm}{Decimal mark used in the currency format}
}
\description{
Converts a string formatted as some currency to a numeric.
}
\examples{
c2numeric('$432.99')

eur <- c('€5000,00', '€432,00', '€100')
c2numeric(eur, dm = ',')
}

