% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{combine.vector}
\alias{combine.vector}
\title{Render column strings for code}
\usage{
combine.vector(x)
}
\value{
Character string of column code
}
\description{
Render column strings for code
}
\keyword{internal}

