% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare}
\usage{
compare(df, group, y)
}
\arguments{
\item{df}{Dataset with variables in question}

\item{group}{string Comparison groups}

\item{y}{string Outcome to compare}
}
\value{
tibble object (1 row) with relevant information
}
\description{
Returns means, standard deviations, and sample sizes for two groups; inferential statistics (t-test and p-value), and effect size with confidence intervals (Cohen's D).
Inputs other than the dataset are strings so that the function can be used with expand_grid variables and map.
}
\examples{
mtcars |> compare("am", "wt")
}
