% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_html.R
\name{display_html}
\alias{display_html}
\title{Display html string rendered in the viewer pane}
\usage{
display_html(str)
}
\arguments{
\item{str}{The html content you want to preview as a string}
}
\value{
Will show the rendered html on the viewer pane
}
\description{
Display html string rendered in the viewer pane
}
