% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harcor.R
\name{harcor}
\alias{harcor}
\title{Title Hot and ready corrleation tables version 2}
\usage{
harcor(data, selection = everything(), controls = NULL, leadzero = F)
}
\arguments{
\item{data}{The dataframe to work with}

\item{selection}{Tidy selection of main variables}

\item{controls}{Tidy selection of control variables}

\item{leadzero}{Leading zero on or off for correlation table}
}
\value{
Nicely formatted correlation table
}
\description{
Title Hot and ready corrleation tables version 2
}
\examples{
mtcars |> harcor(1:9, 10:11)
mmtcars |> harcor(1:9)
}
