% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harcor_to_stargazer.R
\name{harcor_to_stargazer}
\alias{harcor_to_stargazer}
\title{Title Stargazer from harcor V2}
\usage{
harcor_to_stargazer(
  x,
  out = NULL,
  type = "html",
  rownames = F,
  summary = F,
  ...
)
}
\arguments{
\item{x}{harcor output}

\item{out}{file to save to}

\item{type}{choose between 'html','latex', or 'text'. Now only html works appropriately.}

\item{rownames}{T or F. include rownames. defaults to F}

\item{summary}{T or F. F prints the whole dataset as is. Defaults to F}

\item{...}{Optional arguments for stargazer}
}
\value{
stargazer table
}
\description{
Makes a stargazer table from a harcor output
}
