% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/import_scrobbles.R
\name{import_scrobbles}
\alias{import_scrobbles}
\title{Import a Last.fm scrobbles sheet as a data frame.}
\usage{
import_scrobbles(file, header = FALSE, date.fmt = "\%d \%b \%Y \%H:\%M")
}
\arguments{
\item{file}{The name of the file which the scrobble data is to be read from.}

\item{header}{A logical value indicating whether the file contains the names of the variables as its first line.}

\item{date.fmt}{The date format that should be used in the returned data frame.}
}
\value{
A data frame with columns Artist, Album, Song, and Date.
}
\description{
This function initializes a data frame of Last.fm scrobbles.
}
\examples{
import_scrobbles("scrobbles.csv")
}

