% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jekyll.R
\name{knit_htmlwidgets}
\alias{knit_htmlwidgets}
\title{Knit a simulated HTML fragment}
\usage{
knit_htmlwidgets(input, output_dir = "./_includes/htmlwidgets", ...,
  verbose = TRUE)
}
\arguments{
\item{input}{Path of the R Markdown document to be rendered.}

\item{output_dir}{Directory where the rendered HTML should be
written.}

\item{\dots}{Additional arguments to be passed to
\code{rmarkdown::render}.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Invisible TRUE, no matter what.
}
\description{
Renders an R Markdown document as HTML, minus the <!DOCTYPE html>
declaration. Useful as an intermediate step in serving an R Markdown
document with htmlwidget-like content as a Jekyll post.
}
