% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{load_pushbullet}
\alias{load_pushbullet}
\title{Load a Pushbullet profile}
\usage{
load_pushbullet()
}
\description{
To avoid truly attaching RPushbullet, this function simulates all of
the options updates necessary to get the package internals up and
running.
}
\keyword{internal}
