% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{ben.options}
\alias{.options}
\alias{ben.options}
\title{Define package options.}
\usage{
ben.options(...)

.options(..., .__defaults = FALSE, .__reset = FALSE)
}
\arguments{
\item{...}{Either \code{key = value pairs} or the names of the
options to retrieve.}
}
\value{
List of options. If nothing is passed, a list of all options
is returned.
}
\description{
Define package options.
}
\examples{
ben.options()
ben.options('path.alarms')
ben.options(path.alarms = '~')
}

