% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{pbmsg}
\alias{pbmsg}
\title{Send a message via Pushbullet}
\usage{
pbmsg(body = "Your script is done.", title = "Notification from R", ...)
}
\arguments{
\item{body}{Body text of message.}

\item{title}{Title of message.}

\item{\dots}{Additional parameters to be passed to
\code{RPushbullet::pbPost()}.}
}
\description{
Sends a message via Pushbullet. This function is just a lazy wrapper
around \code{RPushbullet::pbPost}.
}
\examples{
\dontrun{pbmsg('Your script has finished.')}
}

