% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_with_csv.R
\name{recode_with_csv}
\alias{recode_with_csv}
\title{Recode factor or character variables with a CSV file
Will take as input a path to a csv file with oldname,newname syntax. Will find each of the oldnames in the column, and replace them for newname.}
\usage{
recode_with_csv(
  data,
  csvfile,
  old_column,
  new_column = {
     {
         old_column
     }
 }
)
}
\arguments{
\item{data}{Dataset to manipulate}

\item{csvfile}{path to a csv file with oldname, newname syntax}

\item{old_column}{variable name to edit}

\item{new_column}{variable name to save the result in. Will overwrite old_column if left blank}
}
\value{
mutated dataset
}
\description{
Recode factor or character variables with a CSV file
Will take as input a path to a csv file with oldname,newname syntax. Will find each of the oldnames in the column, and replace them for newname.
}
\examples{
mtcars |>
mutate(gear_f = factor(gear, labels = c("three", "four","five"))) |>
rename_with_csv(csvfile,gear_f)
}
