% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jekyll.R
\name{remove_doctype}
\alias{remove_doctype}
\title{Strip <!DOCTYPE html> from a file}
\usage{
remove_doctype(input)
}
\arguments{
\item{input}{Path of the HTML file.}
}
\value{
Integer indicating how many lines were removed. Should
  generally only return 1, but the number may be useful for debugging
  unexpected behavior.
}
\description{
Rewrites an HTML file with the <!DOCTYPE html> tag removed.
}
\keyword{internal}
