% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snippets.R
\name{snip.table}
\alias{snip.table}
\title{Render code for a matrix-like object on the clipboard}
\usage{
snip.table(tbl, delim = " ", ...)
}
\arguments{
\item{tbl}{Matrix-like object; if not given, retrieves from the
clipboard}

\item{delim}{The field separator character}

\item{...}{Further arguments to be passed to `readr::read_delim`}
}
\description{
Given some delimited structure on the clipboard or in the session,
this function renders R code to structure the object as a data frame.
The resulting code is written to the clipboard.
}
\examples{
\dontrun{
# Put some table like this on the clipboard:
# mpg cyl disp
# 210 685 87
# 210 610 160
# 710 228 4
snip.table()

# Or like this:
# 210 685 87
# 210 610 160
# 710 228 4
snip.table(col_names = FALSE)

# Or read from a session object:
y <- data.frame(a = c(2,4,6), b = c(4,16,32))
snip.table(y)
}
}

