% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_or.R
\name{stargazer_or}
\alias{stargazer_or}
\title{Stargazer with Odds Ratios}
\usage{
stargazer_or(
  model,
  odd.ratio = T,
  type = "text",
  no.space = T,
  star.cutoffs = c(0.05, 0.01, 0.001),
  ...
)
}
\arguments{
\item{model}{A model object fitted with glm()}

\item{odd.ratio}{logical, should the table contain log-odds coefficients or odds ratios. Defaults to odds ratios.}

\item{...}{Additional parameters for stargazer()}
}
\value{
stargazer table
}
\description{
Stargazer with Odds Ratios
}
\examples{
glm(am ~ wt, data=mtcars,family = 'binomial') |> stargazer_or()
}
