% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{succinct.ws}
\alias{succinct.ws}
\title{Strip extra whitespace used for formatting columns}
\usage{
succinct.ws(x, quote)
}
\value{
Character string
}
\description{
Strip extra whitespace used for formatting columns
}
\keyword{internal}

