% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{time_series_graph}
\alias{time_series_graph}
\title{Create a time series graph of column subsets and their cumulative frequencies.}
\usage{
time_series_graph(df, row.subset, df.col = "Artist", date.col = "Date",
  size = 1)
}
\arguments{
\item{df}{The data frame in which row frequencies are to be calculated. The object should contain a column of dates.}

\item{row.subset}{A subset of df.col values corresponding to the time series graph groups.}

\item{df.col}{The name of the column upon which row frequencies should be calculated.}

\item{date.col}{The name of the column that contains dates.}

\item{size}{A numeric corresponding to line size in the returned plot.}
}
\value{
A ggplot time series graph of frequencies within a column grouped by row.
}
\description{
This function returns a ggplot time series graph of a column, with frequencies grouped by a subset of the column's values.
}
\examples{
time_series_graph(df, c("Artist 1", "Artist 2"))
}

