% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{v2formula}
\alias{v2formula}
\title{Convert a vector of characters to a formula}
\usage{
v2formula(x, y)
}
\arguments{
\item{x}{Vector-like object of explanatory variable names}

\item{y}{Optional character object of the response variable name}
}
\value{
Object of class \code{formula}, like \code{y ~ x1 + x2}
}
\description{
Converts a vector-like object of characters to a formula.
}
\examples{
attrs <- c('x1', 'x2')
v2formula(attrs, 'y')
}

