% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ben.R
\docType{package}
\name{ben}
\alias{ben}
\alias{ben-package}
\title{ben}
\description{
Personal R package of Ben Cunningham.
}
\section{Package Options}{


The following options should be set in \code{.Rprofile} to customize
the behavior of some functions.

\itemize{

 \item \code{ben.alarms}: Directory of sound assets; for example
   \code{~/R/alarms/}

}

To take advantage of the Pushbullet notification wrappers,
\code{RPushbullet} should be configured with
\code{~/.rpushbullet.json} or equivalent global options.
}

