% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{pbmsg}
\alias{pbmsg}
\title{Send a message via Pushbullet}
\usage{
pbmsg(body = "Your script is done.", title = "Notification from R")
}
\arguments{
\item{body}{Body text to be used in message.}

\item{title}{Title to be used in message.}
}
\description{
Sends a message via Pushbullet.
}
\examples{
\dontrun{pbmsg('Your script has finished.')}
}

